questionaire/requirement.txt
flask
fpdf
python-dotenv
google-cloud-aiplatform
